Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@7974 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ randseed/src/HotBitsAndRanArray.java
@@ -24,7 +24,7 @@ public class HotBitsAndRanArray
 	{
 		final URL urlHotBits = new URL("http://www.fourmilab.ch/cgi-bin/uncgi/Hotbits?nbytes="+cSeedBytes+"&fmt=bin");
 
-		final byte[] rUByte = new int[cSeedBytes];
+		final byte[] rUByte = new byte[cSeedBytes];
 
 		InputStream inHotBits = null;
 		try
